+++ crate/agent/src/utils.rs
@@ -1,25 +1,21 @@
 use crate::error::Error;
 use der::{asn1::Ia5String, pem::LineEnding, EncodePem};
-use ima::ima::ImaHashMethod;
 use p256::{ecdsa::DerSignature, ecdsa::SigningKey, pkcs8::EncodePrivateKey, SecretKey};
 use rand_chacha::{
     rand_core::{RngCore, SeedableRng},
     ChaCha20Rng,
 };
-use sha1::{Digest, Sha1};
-use sha2::{Sha256, Sha512};
+
 use spki::{EncodePublicKey, SubjectPublicKeyInfoOwned};
 use std::process::Command;
 use std::{
     convert::TryFrom,
-    fs, io,
     net::{IpAddr, Ipv4Addr},
     path::Path,
     str::FromStr,
     time::Duration,
 };
 use tss_esapi::{Context, TctiNameConf};
-use walkdir::DirEntry;
 use x509_cert::{
     builder::{Builder, CertificateBuilder, Profile},
     ext::pkix::{name::GeneralName, BasicConstraints, SubjectAltName},
@@ -27,57 +23,6 @@ use x509_cert::{
     serial_number::SerialNumber,
     time::Validity,
 };
-
-#[inline(always)]
-pub(crate) fn hash_file(path: &Path, hash_method: &ImaHashMethod) -> Result<Vec<u8>, Error> {
-    let mut file = fs::File::open(path)?;
-
-    match hash_method {
-        ImaHashMethod::Sha1 => {
-            let mut hasher = Sha1::new();
-            let _ = io::copy(&mut file, &mut hasher)?;
-            Ok(hasher.finalize().to_vec())
-        }
-        ImaHashMethod::Sha256 => {
-            let mut hasher = Sha256::new();
-            let _ = io::copy(&mut file, &mut hasher)?;
-            Ok(hasher.finalize().to_vec())
-        }
-        ImaHashMethod::Sha512 => {
-            let mut hasher = Sha512::new();
-            let _ = io::copy(&mut file, &mut hasher)?;
-            Ok(hasher.finalize().to_vec())
-        }
-    }
-}
-
-#[must_use]
-pub fn filter_whilelist(entry: &DirEntry) -> bool {
-    _filter_whilelist(entry).unwrap_or(false)
-}
-
-const BASE_EXCLUDE_DIRS: [&str; 8] = [
-    "/sys/",
-    "/run/",
-    "/proc/",
-    "/lost+found/",
-    "/dev/",
-    "/media/",
-    "/var/",
-    "/tmp/",
-];
-
-pub fn _filter_whilelist(entry: &DirEntry) -> Result<bool, Error> {
-    // Do not keep files in some folders
-    if BASE_EXCLUDE_DIRS
-        .iter()
-        .any(|exclude_dir| entry.path().starts_with(exclude_dir))
-    {
-        return Ok(false);
-    }
-
-    Ok(true)
-}
 
 /// Generate a self-signed certificate
 pub fn generate_self_signed_cert(